+++ src/Views/SnapCorner.cs
@@ -1,0 +1,3 @@
+namespace PrMonitor;
+
+internal enum SnapCorner { None, TopLeft, TopRight, BottomLeft, BottomRight }